Serverhandler Moved to own class
